Dynamic setting of header markup
As the top level header in index.rst and get_started.rst depends on
cookiecutter.service the length of the header markups should also
depend on the length of cookiecutter.service.

Unfortunatelly I was not able to figure out how to include the macro
from a different template, so there is a copy-paste there.

Change-Id: Iff2255f10690f23de270b84c1bd0cd7de3c611a3
Signed-off-by: csatari <gergely.csatari@nokia.com>