Fix directory structure and improve README file
The generated source should be placed in a directory call 'install-guide'
inside existing repositories. Because of that the creation of the repository
directory itself is not possible.

Cookiecutter requires a directory call {{cookiecutter..}} as starting point.
Because of that the only solution without modifing cookiecutter itself is
to add a dummy directory_name variable.

This change also improves the README file. The use of this cookiecutter template
is documeted in the OpenStack Contributor Guide.

Change-Id: I43685feb521a8b0cf53ed4e92dfede8aeddb00b7